% Auto-generated: do not edit by hand
\name{''RGLLayout}

\alias{''RGLLayout}

\title{RGLLayout component}

\description{

}

\usage{
''RGLLayout(children=NULL, id=NULL, layout=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. }

\item{layout}{Unnamed list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

